+++ game/components/c_scene_transition.h
@@ -13,5 +13,6 @@ namespace Verse::Component
     struct SceneTransition {
         Scene* to_scene;
         Vec2 to_pos;
+        str scene_name;
     };
 }