+++ game/controllers/free_actor_controller.h
@@ -7,7 +7,7 @@
 #include "config.h"
 #include "state_machines_list.h"
 
-typedef void (*actor_move_func)(Verse::Config&, Verse::EntityID, Verse::State::StateType);
+typedef bool (*actor_move_func)(Verse::Config&, Verse::EntityID, Verse::State::StateType);
 
 namespace Verse::Controller::FreeActor
 {