+++ game/systems/s_collider.h
@@ -10,6 +10,6 @@ namespace Verse::System::Collider
 {
     std::vector<EntityID> checkObjectCollisions(Config &c, EntityID eid);
     bool checkTilemapCollision(Component::Collider* test_col, Component::Collider* tile_col, Component::Tilemap* tilemap);
-    bool checkCollisions(Config &c, EntityID eid);
+    int checkCollisions(Config &c, EntityID eid);
     void render(Config &c);
 }